New class OpenFoamSimu to post-process simulations with example
